# specs/001-rest-api-refactor/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: College Portal REST API
  version: 1.0.0
  description: REST API contracts for the GraphQL-to-REST migration.
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /registrations:
    get:
      tags: [registrations]
      summary: List registrations (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: studentId
          schema: { type: integer }
        - in: query
          name: semester
          schema: { type: string }
        - in: query
          name: year
          schema: { type: integer }
      responses:
        '200':
          description: Registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Registration'
    post:
      tags: [registrations]
      summary: Register for a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
  /registrations/bulk:
    post:
      tags: [registrations]
      summary: Bulk registration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRegistrationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Registration'
  /registrations/me:
    get:
      tags: [registrations]
      summary: Get my registrations
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: semester
          schema: { type: string }
        - in: query
          name: year
          schema: { type: integer }
      responses:
        '200':
          description: Registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Registration'
  /registrations/me/summary:
    get:
      tags: [registrations]
      summary: Get my registration summary
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: semester
          required: true
          schema: { type: string }
        - in: query
          name: year
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                type: object
  /registrations/{id}:
    get:
      tags: [registrations]
      summary: Get registration by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
    patch:
      tags: [registrations]
      summary: Update registration (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegistrationRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
  /registrations/{id}/payment-status:
    patch:
      tags: [registrations]
      summary: Update payment status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentStatus:
                  type: string
                  enum: [pending, paid, failed, refunded]
              required: [paymentStatus]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
  /registrations/{id}/drop:
    delete:
      tags: [registrations]
      summary: Drop a course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Dropped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /students:
    get:
      tags: [students]
      summary: List students (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
  /students/me:
    get:
      tags: [students]
      summary: Get my student profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    patch:
      tags: [students]
      summary: Update my profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
  /students/me/transcript:
    get:
      tags: [students]
      summary: Get my transcript
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transcript
          content:
            application/json:
              schema:
                type: object
  /students/me/statistics:
    get:
      tags: [students]
      summary: Get my statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema:
                type: object
  /students/me/gpa:
    get:
      tags: [students]
      summary: Get my semester GPA
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: semester
          required: true
          schema: { type: string }
        - in: query
          name: year
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: GPA
          content:
            application/json:
              schema:
                type: object
  /students/{id}:
    get:
      tags: [students]
      summary: Get student by ID (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    patch:
      tags: [students]
      summary: Update student (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    delete:
      tags: [students]
      summary: Delete student (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /students/{id}/transcript:
    get:
      tags: [students]
      summary: Get transcript (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Transcript
          content:
            application/json:
              schema:
                type: object
  /students/{id}/statistics:
    get:
      tags: [students]
      summary: Get statistics (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema:
                type: object
...
  /system-settings:
    get:
      tags: [system-settings]
      summary: List settings (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemSettings'
    patch:
      tags: [system-settings]
      summary: Update a system setting (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemSettingRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
  /system-settings/registration:
    patch:
      tags: [system-settings]
      summary: Toggle registration enabled (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
              required: [enabled]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    Student:
      type: object
      properties:
        id: { type: integer }
        studentId: { type: integer }
        currentGPA: { type: number }
        userId: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        fullName: { type: string }
        isActive: { type: boolean }
        role: { type: string }
...
    Registration:
      type: object
      properties:
        id: { type: integer }
        studentId: { type: integer }
        courseId: { type: integer }
        semester: { type: string }
        year: { type: integer }
        paymentStatus: { type: string, enum: [pending, paid, failed, refunded] }
        grade: { type: string, enum: [A, B, C, D, F, I, W], nullable: true }
        gradePoints: { type: number, nullable: true }
        isCompleted: { type: boolean }
        isDropped: { type: boolean }
        droppedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    SystemSettings:
      type: object
      properties:
        id: { type: integer }
        settingKey: { type: string }
        settingValue: { type: string }
        description: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateRegistrationRequest:
      type: object
      properties:
        studentId: { type: integer }
        courseId: { type: integer }
        semester: { type: string }
        year: { type: integer }
      required: [courseId, semester, year]
    BulkRegistrationRequest:
      type: object
      properties:
        studentId: { type: integer }
        courseIds:
          type: array
          items: { type: integer }
        semester: { type: string }
        year: { type: integer }
      required: [courseIds, semester, year]
    UpdateRegistrationRequest:
      type: object
      properties:
        paymentStatus: { type: string }
        grade: { type: string }
        isCompleted: { type: boolean }
        isDropped: { type: boolean }
        droppedAt: { type: string, format: date-time }
...
    UpdateStudentRequest:
      type: object
      properties:
        currentGPA: { type: number }
        isActive: { type: boolean }
    UpdateSystemSettingRequest:
      type: object
      properties:
        settingKey: { type: string }
        settingValue: { type: string }
      required: [settingKey, settingValue]